backend/services/paddle_ocr/requirements.txt
paddlepaddle==2.6.2
paddleocr==2.7.0.3
fastapi==0.110.0
uvicorn[standard]==0.29.0
Pillow==10.2.0


